+++ jms/appServer/static/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 1.3.6
 -----
 Added a LocalConnectionFactory for ActiveMQ
+
 1.3.5
 -----
 Added the ability to declare custom JVM System Properties in your stanzas