+++ jmx/appServer/static/docs/readme/RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+2.1
+----
+Config file dynamically reloaded if it changes
+PID File contents read in on each poller execution
+PID Command execution on each poller execution
+PID Command can also return JVM Descriptions
+
 2.0.4
 -----
 Minor change so that when using "dumpAllAttributes" , only READABLE attributes will be polled.